Properly update profile connection state.
When DUT is toggled between various paired Headsets Connection
state is not updated properly in some scenarios.This is causing
Bluetooth icon to be in connected state even when no Headsets are
connected.

With this patch we will broadcast connection states in case of
failure and will make sure connected, connecting,disconnecting
devices count is never negative.

Change-Id: Icb1a7215ff6d9491afbb5f7f203ca0a2c6b7f43a

Conflicts:
	src/com/android/bluetooth/a2dp/A2dpStateMachine.java